Add token-service application (#10)
Towards a client appliation, the token-service application acts
as an identity provider. It lets a client use various types
of OAuth client credentials flows to authenticate, and returns
an ID token that asserts the client's identity.

Initially, the application only supports a single authentication
flow, xlb-mtls-client-credentials. This flow authenticates clients
using mTLS and relies on the Google Cloud load balancer's mTLS feature.
The application is designed so that other flows can be added
and enabled/disabled individually.

When registered as a workload identity pool provider, ID tokens
issued by the token-service application can be exchanged against
short-lived Google credentials by using the Google STS.

The application uses Quarkus, CDI, and Jax-RS and is designed
to run on Cloud Run.

Review: 19060
Change-Id: I43276f39b40b24e29e12f5f7e3c2f89a68f110a1